Add Monitor.Alerts
Add the ability to pass a single list of alerts to a monitor that will
be executed on either an up or down alert.

Right now this can be used with a Go template that changes the alert
content depending on the type of alert. Later, I may add a way to pass
two commands to an Alert where one can be for down and the other for
up.